+++ package.json
@@ -28,6 +28,7 @@
     "debounce": "^1.0.0",
     "deep-equal": "^1.0.1",
     "dlnacasts": "^0.1.0",
+    "download-file": "^0.1.5",
     "drag-drop": "^4.1.0",
     "electron": "^4.0.0",
     "es6-error": "^4.0.0",
@@ -40,11 +41,13 @@
     "mkdirp": "^0.5.1",
     "music-metadata": "^3.3.0",
     "network-address": "^1.1.0",
+    "node-gzip": "^1.1.2",
     "parse-torrent": "^6.0.1",
     "prettier-bytes": "^1.0.1",
     "prop-types": "^15.6.2",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
+    "request": "^2.88.0",
     "request-promise-native": "^1.0.5",
     "rimraf": "^2.5.2",
     "run-parallel": "^1.1.6",
